pyproject.toml
[project]
name = "telehealth-agentic-ai-app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "boto3>=1.34.0",
    "strands-agents>=0.1.0",
    "faiss-cpu>=1.8.0",
    "numpy>=1.24.0",
    "asyncio>=3.4.3",
    "uuid>=1.30",
    "logging>=0.4.9.6",
    "streamlit"
]
